wip: more ports, all audio block and programs 1-4

+++ Inc/audio/bitcrusher.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 typedef struct 
 {
-    uint16_t bitmask;
+    uint32_t bitmask;
 } BitCrusherDataType;
