+++ Inc/audio/reverb2.h
@@ -10,13 +10,13 @@ typedef struct
     AllpassType aps[8];
     DelayDataType delaylines[4];
     FirstOrderIirType lowpass;
-    int16_t mix;
-    int16_t decay;
-    int16_t outs[4];
+    float mix;
+    float decay;
+    float outs[4];
 } Reverb2Type;
 
 
-int16_t reverb2ProcessSample(int16_t sampleIn,Reverb2Type*reverbData);
+float reverb2ProcessSample(float sampleIn,Reverb2Type*reverbData);
 void initReverb2(Reverb2Type*reverbData);
 
 #endif
